requirements.txt
biopython>=1.83
faiss-cpu>=1.8.0
goatools>=1.4.12
matplotlib>=3.8.0
numpy>=1.26.0
pandas>=2.2.0
pyyaml>=6.0.1
requests>=2.31.0
scikit-learn>=1.4.0
seaborn>=0.13.0
setuptools>=68.0.0
torch>=2.2.0
tqdm>=4.66.0
transformers>=4.39.0
